apps/api/pyproject.toml
[project]
name = "perdiz-api"
version = "0.0.0"
description = "p3rDiz — backend FastAPI"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "python-jose[cryptography]>=3.3",
  "argon2-cffi>=23.1",
  "authlib>=1.3",
  "httpx>=0.27",
  "structlog>=24.1",
  "sentry-sdk[fastapi]>=1.45",
  "slowapi>=0.1.9",
  "redis>=5.0",
  "arq>=0.25",
  "boto3>=1.34",
  "python-ulid>=2.2",
  "bleach>=6.1",
  "trimesh>=4.3",
  "resend>=0.7",
  "email-validator>=2.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "httpx>=0.27",
  "testcontainers[postgres]>=4.3",
  "respx>=0.21",
  "fakeredis>=2.21",
  "ruff>=0.4",
  "mypy>=1.9",
  "types-bleach",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "PL", "RUF"]
ignore = ["PLR0913", "PLR2004"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PLR2004", "B011"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_generics = true
plugins = []

[[tool.mypy.overrides]]
module = ["trimesh.*", "resend.*", "arq.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
